+++ app/src/main/java/com/malakezzat/banquemisr/challenge05/data/local/MoviesLocalDataSource.kt
@@ -11,6 +11,6 @@ interface MoviesLocalDataSource {
     suspend fun getUpcomingMovies(): Flow<List<MovieDB>>?
 
     suspend fun insertMovieDetails(movieDetailsDB: MovieDetailsDB)
-    suspend fun getMovieDetailsById(id: Int): Flow<MovieDetailsDB>?
+    suspend fun getMovieDetailsById(id: Long): Flow<MovieDetailsDB>?
 }
